chore: Make use of the new tx_sender helper test functions.
